[FIX] adjust date_start and date_end of sale.order.line for contract
1. set invisible of field date_start and date_end
Because we have already changed the field date_start and date_end to related fields.
So they are readonly in view now. The user should set the start_date and end_date
instead of date_start and date_end. In order to set the start_date and end_date for
service product (is_contract but not rental). The Flag 'must_have_dates' of this product should be set.

2. add onchange function to calculate product_uom_qty
If user add a sale order line for product (is_contract=True and not rental).
the product_uom_qty should be calculated with the set dates.